+++ app/src/main/java/com/example/fe/scrap/ScrapBottomSheetFragment.kt
@@ -7,7 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.fe.R
 import com.example.fe.bookclub_place.api.RetrofitClient
 import com.example.fe.databinding.FragmentScrapBottomSheetBinding
 import com.example.fe.mypage.ScrapFolderResponse
@@ -17,7 +16,7 @@ import retrofit2.Callback
 import retrofit2.Response
 
 class ScrapBottomSheetFragment(
-    private val bookId: Int?,
+    private val bookId: Long?,
     private val placeId: Int?,
     private val onBookmarkStateChanged: (Boolean) -> Unit // 선택/해제 상태 콜백 추가
 ) : BottomSheetDialogFragment() {
